requirements.txt
pytest==8.3.3

mysql-connector-python~=9.1.0